backend/requirements.txt
langgraph
langchain
langchain-google-genai
python-dotenv
apify-client
numpy
isodate
google-api-python-client
pydantic
scikit-learn